add controller , tomorrow gonna do the last step , add Docker

+++ rinhabackend-java/src/main/java/br/com/jujubaprojects/rinhabackendjava/dto/TransacaoResponse.java
@@ -6,7 +6,7 @@ public class TransacaoResponse {
 
     private Character tipo;
 
-    private int saldo;
+    private double saldo;
